chore: organize codebase naming and import orders
