+++ src/components/layouts/FactLayout.jsx
@@ -1,5 +1,6 @@
-import "../App.css";
-export const Fact = (props) => {
+import "../../App.css";
+
+export const FactLayout = (props) => {
   return (
     <div className="fact flex-column animate__animated animate__fadeInLeft ">
       <img src={props.image} alt="the planet image" />
