+++ src/pages/Explore.jsx
@@ -1,4 +1,5 @@
 import { Link } from "react-router-dom";
+
 import { Orion, IvaSuit } from "../assets";
 const ExplorePage = () => {
   return (
