+++ src/pages/Mars.jsx
@@ -1,11 +1,13 @@
-import { Planets } from "../components/planet";
-import { Fact } from "../components/fact";
 import { Link } from "react-router-dom";
+
+import { PlanetLayout } from "../components/layouts/PlanetLayout";
+import { FactLayout } from "../components/layouts/FactLayout";
+
 import { Mars, OlympusMons, MarsValleries, CuriosityRover } from "../assets";
 
 const MarsPage = () => {
   return (
-    <Planets id="2">
+    <PlanetLayout id="2">
       <section>
         {/* animating elements from the left overflows the website */}
         <p className="greet animate__animated animate__bounceInLeft">
@@ -20,15 +22,15 @@ const MarsPage = () => {
           <h1 className="animate__animated animate__bounceInLeft">Mars</h1>
         </div>
 
-        <Fact
+        <FactLayout
           image={OlympusMons}
           content="Behold Olympus Mons, the tallest volcano in the solar system! It's so massive you could host the ultimate interplanetary ski competition on its slopes. "
         />
-        <Fact
+        <FactLayout
           image={MarsValleries}
           content="Zoom in on Valles Marineris, a canyon system that makes the Grand Canyon look like a crack in the sidewalk. Mars goes big or goes home! "
         />
-        <Fact
+        <FactLayout
           image={CuriosityRover}
           content="Fun fact - our trusty rover friend, Curiosity, is roaming around down there. It's like the solar system's own remote-controlled explorer. "
         />
@@ -46,7 +48,7 @@ const MarsPage = () => {
           </button>
         </Link>
       </section>
-    </Planets>
+    </PlanetLayout>
   );
 };
 export default MarsPage;
